# api/openapi.yaml
openapi: 3.0.0
info:
  title: EA Trading api API
  description: API for managing orders, holdings, and client portfolios in the EA Trading system
  version: 1.0.0
  contact:
    name: EA Trading Support
    email: support@eatrading.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://api.eatrading.com
    description: Production server

tags:
  - name: Clients
    description: Client management operations
  - name: Orders
    description: Order operations (buy/sell transactions)
  - name: Holdings
    description: Client portfolio holdings operations

paths:
  /api/clients:
    get:
      tags:
        - Clients
      summary: Get clients
      description: |
        Retrieve clients. 
        - Without clientId parameter: Returns a list of all clients in the system
        - With clientId parameter: Returns a specific client by their ID (users can only access their own client data)
      operationId: getClients
      parameters:
        - name: clientId
          in: query
          required: false
          description: The UUID of the client. If provided, returns a specific client; if omitted, returns all clients.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Successfully retrieved client(s)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Client'
                  - $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Clients
      summary: Update client information
      description: Update a client's name and email. Users can only update their own profile.
      operationId: updateClient
      parameters:
        - name: clientId
          in: query
          required: true
          description: The UUID of the client
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        description: Client update information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdateRequest'
      responses:
        '200':
          description: Client successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/orders:
    get:
      tags:
        - Orders
      summary: Get orders for a client
      description: Retrieve all orders for a specific client. Users can only view their own orders.
      operationId: getOrdersByClient
      parameters:
        - name: clientId
          in: query
          required: true
          description: The UUID of the client
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Orders successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/orders/transact:
    post:
      tags:
        - Orders
      summary: Create a new transaction order
      description: Create a new BUY or SELL transaction order. Users can only create orders for their own account.
      operationId: createTransactionOrder
      requestBody:
        required: true
        description: Order transaction details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderTransactionRequest'
      responses:
        '201':
          description: Order successfully created and submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/holdings:
    get:
      tags:
        - Holdings
      summary: Get client holdings
      description: Retrieve all holdings in a client's portfolio. Users can only view their own holdings.
      operationId: getClientHoldings
      parameters:
        - name: clientId
          in: query
          required: true
          description: The UUID of the client
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Holdings successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HoldingResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    Client:
      type: object
      description: Client entity representing a trading client
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the client
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Client's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: Client's email address
          example: "john.doe@example.com"
        portfolio:
          $ref: '#/components/schemas/Portfolio'
        watchlist:
          type: array
          description: List of assets in the client's watchlist
          items:
            $ref: '#/components/schemas/Asset'
      required:
        - id
        - name
        - email

    ClientUpdateRequest:
      type: object
      description: Request body for updating client information
      properties:
        name:
          type: string
          description: Client's full name
          example: "John Doe"
          minLength: 1
        email:
          type: string
          format: email
          description: Client's email address
          example: "john.doe@example.com"
      required:
        - name
        - email

    Order:
      type: object
      description: Order entity representing a buy or sell transaction
      properties:
        orderId:
          type: string
          format: uuid
          description: Unique identifier for the order
          example: "660e8400-e29b-41d4-a716-446655440001"
        clientId:
          type: string
          format: uuid
          description: UUID of the client who placed the order
          example: "550e8400-e29b-41d4-a716-446655440000"
        buy:
          type: boolean
          description: True if BUY order, false if SELL order
          example: true
        orderDate:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the order was placed
          example: "2024-09-25T14:30:00Z"
        price:
          type: number
          format: double
          description: Price per unit at the time of order
          example: 150.50
        quantity:
          type: number
          format: double
          description: Number of units in the order
          example: 10.50
        asset:
          $ref: '#/components/schemas/Asset'
        status:
          type: string
          enum:
            - SUBMITTED
            - PENDING
            - EXECUTED
            - CANCELLED
            - REJECTED
          description: Current status of the order
          example: "SUBMITTED"
      required:
        - orderId
        - clientId
        - buy
        - orderDate
        - price
        - quantity
        - status

    OrderTransactionRequest:
      type: object
      description: Request body for creating a new transaction order
      properties:
        clientId:
          type: string
          format: uuid
          description: UUID of the client placing the order
          example: "550e8400-e29b-41d4-a716-446655440000"
        symbol:
          type: string
          description: Trading symbol or ticker for the asset
          example: "AAPL"
        transactionType:
          type: string
          enum:
            - BUY
            - SELL
          description: Type of transaction
          example: "BUY"
        quantity:
          type: number
          format: double
          description: Number of units to trade
          example: 10.50
          minimum: 0
          exclusiveMinimum: true
      required:
        - clientId
        - symbol
        - transactionType
        - quantity

    OrderTransactionResponse:
      type: object
      description: Response body after creating a transaction order
      properties:
        orderId:
          type: string
          format: uuid
          description: Unique identifier for the created order
          example: "660e8400-e29b-41d4-a716-446655440001"
        clientId:
          type: string
          format: uuid
          description: UUID of the client who placed the order
          example: "550e8400-e29b-41d4-a716-446655440000"
        symbol:
          type: string
          description: Trading symbol for the asset
          example: "AAPL"
        transactionType:
          type: string
          enum:
            - BUY
            - SELL
          description: Type of transaction
          example: "BUY"
        quantity:
          type: string
          description: Number of units in the order
          example: "10.50"
        status:
          type: string
          enum:
            - SUBMITTED
            - PENDING
            - EXECUTED
            - CANCELLED
            - REJECTED
          description: Current status of the order
          example: "SUBMITTED"
        orderDate:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the order was created
          example: "2024-09-25T14:30:00Z"
      required:
        - orderId
        - clientId
        - symbol
        - transactionType
        - quantity
        - status
        - orderDate

    HoldingResponse:
      type: object
      description: Response containing a client's asset holding information
      properties:
        clientId:
          type: string
          format: uuid
          description: UUID of the client
          example: "550e8400-e29b-41d4-a716-446655440000"
        symbol:
          type: string
          description: Trading symbol for the asset
          example: "AAPL"
        quantity:
          type: string
          description: Total quantity held
          example: "100.50"
        avgBuyPrice:
          type: string
          description: Average price paid for the holdings
          example: "145.25"
        currentValue:
          type: string
          description: Current market value of the holdings
          example: "15525.50"
      required:
        - clientId
        - symbol
        - quantity
        - avgBuyPrice
        - currentValue

    Portfolio:
      type: object
      description: Portfolio entity containing all holdings for a client
      properties:
        portfolioId:
          type: string
          format: uuid
          description: Unique identifier for the portfolio
          example: "770e8400-e29b-41d4-a716-446655440002"
        clientId:
          type: string
          format: uuid
          description: UUID of the client who owns the portfolio
          example: "550e8400-e29b-41d4-a716-446655440000"
        holdings:
          type: array
          description: List of asset holdings in the portfolio
          items:
            $ref: '#/components/schemas/Holding'
        totalValue:
          type: number
          format: double
          description: Total market value of the portfolio
          example: 50000.00
      required:
        - portfolioId
        - clientId

    Holding:
      type: object
      description: Asset holding in a portfolio
      properties:
        holdingId:
          type: string
          format: uuid
          description: Unique identifier for the holding
          example: "880e8400-e29b-41d4-a716-446655440003"
        asset:
          $ref: '#/components/schemas/Asset'
        quantity:
          type: number
          format: double
          description: Total quantity held
          example: 100.50
        averageBuyPrice:
          type: number
          format: double
          description: Average price paid per unit
          example: 145.25
      required:
        - holdingId
        - asset
        - quantity

    Asset:
      type: object
      description: Financial asset (stock, commodity, etc.)
      properties:
        assetId:
          type: string
          format: uuid
          description: Unique identifier for the asset
          example: "990e8400-e29b-41d4-a716-446655440004"
        symbol:
          type: string
          description: Trading symbol or ticker
          example: "AAPL"
        name:
          type: string
          description: Full name of the asset
          example: "Apple Inc."
        assetType:
          type: string
          description: Type of asset (STOCK, COMMODITY, etc.)
          example: "STOCK"
        currentPrice:
          type: number
          format: double
          description: Current market price
          example: 154.50
      required:
        - assetId
        - symbol
        - name
        - assetType
        - currentPrice

    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-09-25T14:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type or title
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message
          example: "Invalid client ID format"
        path:
          type: string
          description: The API path that caused the error
          example: "/api/clients"
      required:
        - timestamp
        - status
        - error
        - message

  responses:
    BadRequestError:
      description: Bad request - Invalid input or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 400
            error: "Bad Request"
            message: "Invalid request format or missing required fields"
            path: "/api/orders/transact"

    UnauthorizedError:
      description: Unauthorized - Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 401
            error: "Unauthorized"
            message: "Authentication required"
            path: "/api/clients"

    ForbiddenError:
      description: Forbidden - User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 403
            error: "Forbidden"
            message: "You do not have permission to access this resource"
            path: "/api/clients"

    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 404
            error: "Not Found"
            message: "Client not found"
            path: "/api/clients"

    UnprocessableEntityError:
      description: Unprocessable entity - Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 422
            error: "Unprocessable Entity"
            message: "Email validation failed"
            path: "/api/clients"

    InternalServerError:
      description: Internal server error - Server encountered an unexpected condition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2024-09-25T14:30:00Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            path: "/api/orders/transact"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer Token Authentication. 
        All endpoints require a valid JWT token in the Authorization header.
        
        Format: Authorization: Bearer jwt
  
security:
  - BearerAuth: []